flask-1/requirements.txt
blinker-1.4
certifi-2020.12.5
chardet-4.0.0
click-7.1.2
Flask_DebugToolbar-0.11.0
Flask-1.1.2
forex_python-1.5
idna-2.10
itsdangerous-1.1.0
jinja2-2.11.3
Markupsafe-1.1.1
werkzeug-1.0.1
